Updated onslaught with channel checker

+++ cogs/moderation.py
@@ -8,20 +8,24 @@ class ModerationCog(commands.Cog, name='Moderation'):
 
     @commands.command()
     async def purge(self, ctx, number:int=None):
-        print(">>>purge command called")
-        if ctx.message.author.guild_permissions.manage_messages:
-            try:
-                if number is None:
-                    await ctx.send('You must input a number')
-                else:
-                    await ctx.message.delete()
-                    
-                    deleted = await ctx.message.channel.purge(limit=number)
-                    await ctx.send(f'Messages purged by {ctx.message.author.mention}: {len(deleted)}')
-            except:
-                await ctx.send("I can't purge messages here, sorry.")
+        enabled = False
+        if enabled:
+            print(">>>purge command called")
+            if ctx.message.author.guild_permissions.manage_messages:
+                try:
+                    if number is None:
+                        await ctx.send('You must input a number')
+                    else:
+                        await ctx.message.delete()
+                        
+                        deleted = await ctx.message.channel.purge(limit=number)
+                        await ctx.send(f'Messages purged by {ctx.message.author.mention}: {len(deleted)}')
+                except:
+                    await ctx.send("I can't purge messages here, sorry.")
+            else:
+                await ctx.send("You do not have permissions to use this command!")
         else:
-            await ctx.send("You do not have permissions to use this command!")
+            await ctx.send("Command is disabled, have a chit chat with Relsie if you want this mf thang")
 
 async def setup(bot):
     await bot.add_cog(ModerationCog(bot))
